refactor: remove dead code identified in the 2026-06-10 review
Deletions (no callers anywhere in package, tests, or experiments;
verified by grep and a full-package import sweep):
- graphik/utils/operators.py (entire module) and the dgp.py helpers
  that only it consumed (orthogonal_procrustes, factor, sample_matrix,
  incidence_matrix_from_adjacency)
- graphik/utils/chordal.py: verbatim copy of networkx's
  complete_to_chordal_graph; sdp_snl now calls nx's directly
- graphik/solvers/constraints.py: SOS-era sympy layer, broken since the
  2D/3D unification (robot_graph.directed, parent_node_id). The one
  live function, get_full_revolute_nearest_point, moved to sdp_snl.py
- utils.py: dZ, perpendicular_vector, norm_sq, list_to_variable_dict_-
  spherical, variable_dict_to_list, make_save_string, spherical_angle_-
  bounds_to_revolute, safe_arccos, bernoulli/wilson confidence helpers,
  measure_perturbation, constraint_violations, __main__ block
- kinematics.py: fk_2d (only self-recursive)
- robot.py: end_effector_pos (iterated characters of a node name),
  jacobian_geometric (referenced removed .parents), lambdified,
  spherical
- graph.py: distance_bounds_from_sampling (dead, wrote garbage DIST)
- joint_angle_solver.py: gen_objective_ee/gen_grad_ee (superseded by
  gen_cost_and_grad_ee), main() with ~60 lines of commented-out demo,
  unused imports
- sdp_snl.py: sym_vec, constraints_list_to_matrix, ~130-line __main__
  Mosek demo, unused typ assignment
- convex_iteration.py: random_psd_matrix, solve_fantope_sdp (+ its
  only-callers fantope_constraints, solve_fantope_iterate),
  get_sparsity_pattern, ~150-line __main__ benchmark
- robot_visualization.py: import-time rc('text', usetex=True) side
  effect (broke import without LaTeX) is now an opt-in use_latex_text()

setup.py: sympy and progress dependencies dropped (only dead code used
them). Suite: 91 passed.
